[ADD] feed view 요청 파라미터 수정

+++ src/main/java/stackpot/stackpot/web/dto/FeedResponseDto.java
@@ -16,7 +16,7 @@ public class FeedResponseDto {
     @NoArgsConstructor
     public static class FeedPreviewList {
         private List<FeedDto> feeds;
-        private Integer nextCursor; // 다음 커서 값
+        private Long nextCursor; // 다음 커서 값
     }
 
     @Data
